Update Project 6
This is just some super basic first ideas. I ended up having to initialize the array in main because of troubles with scope of the hotplate array. In all my testing so far this looks good but feel free to check on what I've done so far and all of this can be subject to change as we go on.